// src/app/models/collecte-donnees.ts
export interface CollecteDonnees {
  idCollecte: number;
    dateCollecte: string;
    produit: string;
    prixUnitaire: number;
    quantiteDisponible: number;
    idMarche: number;
    nomMarche: string;
    lieuMarche: string;
    idEnqueteur: number;
    nomEnqueteur: string;
    prenomEnqueteur: string;
}